Prototype simple tool to gather data (#3)
A simple command-line tool has been implemented that goes through a
given number of days and tries to download a specific artifact for each
day. If the artifact is already cached in either Fastly or CloudFront,
it will be skipped and the tool will continue with the previous day.
In the end, a list with the download speeds of all samples will be
printed.